requirements.txt
telethon
fasttelethon
ffmpeg-python
tqdm
